requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.28.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
argparse=1.4.0=pypi_0
asyncio=3.4.3=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py312h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-runtime=12.1.0=0
dataclasses-json=0.6.4=pypi_0
distro=1.9.0=pypi_0
einops=0.7.0=pypi_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
greenlet=3.0.3=pypi_0
h11=0.14.0=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.1=pypi_0
huggingface-hub=0.22.2=pypi_0
idna=3.4=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py312h06a4308_0
joblib=1.2.0=py312h06a4308_0
jpeg=9e=h5eee18b_1
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
lame=3.100=h7b6447c_0
langchain=0.1.13=pypi_0
langchain-community=0.0.29=pypi_0
langchain-core=0.1.33=pypi_0
langchain-experimental=0.0.55=pypi_0
langchain-text-splitters=0.0.1=pypi_0
langsmith=0.1.49=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
loguru=0.7.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py312h5eee18b_0
marshmallow=3.21.1=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
mpmath=1.3.0=py312h06a4308_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py312h06a4308_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
openai=1.23.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
orjson=3.10.1=pypi_0
packaging=23.2=pypi_0
pandas=2.2.2=pypi_0
pillow=10.2.0=py312h5eee18b_0
pip=23.3.1=py312h06a4308_0
psutil=5.9.8=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pygments=2.17.2=pypi_0
pypdf=4.1.0=pypi_0
python=3.12.3=h996f2a0_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytorch=2.2.2=py3.12_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py312h5eee18b_0
ratelimit=2.2.1=pypi_0
readline=8.2=h5eee18b_0
regex=2024.4.16=pypi_0
requests=2.31.0=py312h06a4308_1
rich=13.5.2=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.4.2=py312h526ad5a_1
scipy=1.12.0=py312hc5e2394_0
sentry-sdk=1.45.0=pypi_0
setuptools=68.2.2=py312h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.29=pypi_0
sqlite=3.41.2=h5eee18b_0
swarms=4.8.2=pypi_0
sympy=1.12=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.19.1=pypi_0
toml=0.10.2=pypi_0
torchaudio=2.2.2=py312_cu121
torchvision=0.17.2=py312_cu121
tqdm=4.66.2=pypi_0
transformers=4.40.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.9.0=py312h06a4308_1
tzdata=2024.1=pypi_0
ucimlrepo=0.0.6=pypi_0
urllib3=2.1.0=py312h06a4308_0
wheel=0.41.2=py312h06a4308_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
